Use fmt::print() to print data from pipe
The child writes text to the pipe encoded as UTF-8.
We use fmt::print() instead of fwrite() to write to stdout, because
stdout likely point to the console and the code page of the console
might not be UTF-8. If we use fmt::print(), we can pass UTF-8 text to
it, because fmt::print() will use the appropriate console APIs for
Unicode.