stdin-echo: Set Console Mode of Input
Set mode so that ReadConsoleW returns one character at a time.

Furthermore Search for Ctrl-C, Ctrl-D, and CR in input.

Search for ^C becuase an Control Event is not generated, becuase the
mode was changed.

Search for ^D to work like Linux. ^D is EOF by default.

The enter key yields CR. So we replace it with LF, like the option
"icrlf" of stty on linux, which is enabled by default.