Add a shared items project for helper.{cpp,h}
And add an additional include path: That is the folder "include" in the
root of the git repository.  Inside of the "include" folder is the
folder "console-tools" (full path "include/console-tools").  So a header
inside that folder must be included like this:

  #include "console-tools/header.h"

That makes it obvious, that the header is from the current project.